# src/typescript/shared/openapi/hevy/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hevy API Docs",
    "description": "\n      Welcome to Hevy's public API! We're just starting to roll this out and depending on your\n      feedback, we'll be adding more features and endpoints. Also, we make no guarantees that \n      we won't completely change the structure or abandon the project entirely so use it at\n      your own risk. Currently, this API is only available to Hevy Pro users. You can get your \n      key on our web app at https://hevy.com/settings?developer. If you have any questions,\n      please reach out to pedro@hevyapp.com 🤝.\n      ",
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/workouts/{workoutId}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get a single workout’s complete details by the workoutId",
        "parameters": [
          {
            "in": "header",
            "name": "api-key",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "workoutId",
            "in": "path",
            "description": "The id of the workout",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          },
          "404": {
            "description": "Workout not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing workout",
        "tags": [
          "Workouts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "api-key",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "workoutId",
            "description": "The id of the workout",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostWorkoutsRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The workout was successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workout"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PostWorkoutsRequestSet": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the set.",
            "enum": [
              "warmup",
              "normal",
              "failure",
              "dropset"
            ],
            "example": "normal"
          },
          "weight_kg": {
            "type": "number",
            "nullable": true,
            "description": "The weight in kilograms.",
            "example": 100
          },
          "reps": {
            "type": "integer",
            "nullable": true,
            "description": "The number of repetitions.",
            "example": 10
          },
          "distance_meters": {
            "type": "integer",
            "nullable": true,
            "description": "The distance in meters.",
            "example": null
          },
          "duration_seconds": {
            "type": "integer",
            "nullable": true,
            "description": "The duration in seconds.",
            "example": null
          },
          "custom_metric": {
            "type": "number",
            "nullable": true,
            "description": "A custom metric for the set. Currently used for steps and floors.",
            "example": null
          },
          "rpe": {
            "type": "number",
            "nullable": true,
            "description": "The Rating of Perceived Exertion (RPE).",
            "enum": [
              6,
              7,
              7.5,
              8,
              8.5,
              9,
              9.5,
              10
            ],
            "example": null
          }
        }
      },
      "PostWorkoutsRequestExercise": {
        "type": "object",
        "properties": {
          "exercise_template_id": {
            "type": "string",
            "description": "The ID of the exercise template.",
            "example": "D04AC939"
          },
          "superset_id": {
            "type": "integer",
            "nullable": true,
            "description": "The ID of the superset.",
            "example": null
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes for the exercise.",
            "example": "Felt good today. Form was on point."
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostWorkoutsRequestSet"
            }
          }
        }
      },
      "PostWorkoutsRequestBody": {
        "type": "object",
        "properties": {
          "workout": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "description": "The title of the workout.",
                "example": "Friday Leg Day 🔥"
              },
              "description": {
                "type": "string",
                "nullable": true,
                "description": "A description for the workout workout.",
                "example": "Medium intensity leg day focusing on quads."
              },
              "start_time": {
                "type": "string",
                "description": "The time the workout started.",
                "example": "2024-08-14T12:00:00Z"
              },
              "end_time": {
                "type": "string",
                "description": "The time the workout ended.",
                "example": "2024-08-14T12:30:00Z"
              },
              "is_private": {
                "type": "boolean",
                "description": "A boolean indicating if the workout is private.",
                "example": false
              },
              "exercises": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PostWorkoutsRequestExercise"
                }
              }
            }
          }
        }
      },
...
      "Workout": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The workout ID.",
            "example": "b459cba5-cd6d-463c-abd6-54f8eafcadcb"
          },
          "title": {
            "type": "string",
            "description": "The workout title.",
            "example": "Morning Workout 💪"
          },
          "routine_id": {
            "type": "string",
            "description": "The ID of the routine that this workout belongs to.",
            "example": "b459cba5-cd6d-463c-abd6-54f8eafcadcb"
          },
          "description": {
            "type": "string",
            "description": "The workout description.",
            "example": "Pushed myself to the limit today!"
          },
          "start_time": {
            "type": "string",
            "description": "ISO 8601 timestamp of when the workout was recorded to have started.",
            "example": "2021-09-14T12:00:00Z"
          },
          "end_time": {
            "type": "string",
            "description": "ISO 8601 timestamp of when the workout was recorded to have ended.",
            "example": "2021-09-14T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "description": "ISO 8601 timestamp of when the workout was last updated.",
            "example": "2021-09-14T12:00:00Z"
          },
          "created_at": {
            "type": "string",
            "description": "ISO 8601 timestamp of when the workout was created.",
            "example": "2021-09-14T12:00:00Z"
          },
          "exercises": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "number",
                  "description": "Index indicating the order of the exercise in the workout.",
                  "example": 0
                },
                "title": {
                  "type": "string",
                  "description": "Title of the exercise",
                  "example": "Bench Press (Barbell)"
                },
                "notes": {
                  "type": "string",
                  "description": "Notes on the exercise",
                  "example": "Paid closer attention to form today. Felt great!"
                },
                "exercise_template_id": {
                  "type": "string",
                  "description": "The id of the exercise template. This can be used to fetch the exercise template.",
                  "example": "05293BCA"
                },
                "supersets_id": {
                  "type": "number",
                  "nullable": true,
                  "description": "The id of the superset that the exercise belongs to. A value of null indicates the exercise is not part of a superset.",
                  "example": 0
                },
                "sets": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "index": {
                        "type": "number",
                        "description": "Index indicating the order of the set in the workout.",
                        "example": 0
                      },
                      "type": {
                        "type": "string",
                        "description": "The type of set. This can be one of 'normal', 'warmup', 'dropset', 'failure'",
                        "example": "normal"
                      },
                      "weight_kg": {
                        "type": "number",
                        "nullable": true,
                        "description": "Weight lifted in kilograms.",
                        "example": 100
                      },
                      "reps": {
                        "type": "number",
                        "nullable": true,
                        "description": "Number of reps logged for the set",
                        "example": 10
                      },
                      "distance_meters": {
                        "type": "number",
                        "nullable": true,
                        "description": "Number of meters logged for the set",
                        "example": null
                      },
                      "duration_seconds": {
                        "type": "number",
                        "nullable": true,
                        "description": "Number of seconds logged for the set",
                        "example": null
                      },
                      "rpe": {
                        "type": "number",
                        "nullable": true,
                        "description": "RPE (Relative perceived exertion) value logged for the set",
                        "example": 9.5
                      },
                      "custom_metric": {
                        "type": "number",
                        "nullable": true,
                        "description": "Custom metric logged for the set (Currently only used to log floors or steps for stair machine exercises)",
                        "example": 50
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}